+++ docs/CHANGELOG.md
@@ -13,6 +13,8 @@ and this programmer does her best to follow [Conventional Commits](https://www.c
 
 ### Changed
 
+- Moved PerfLogger instance listing utils to newly created Logger base class
+
 ### Removed
 
 ### Fixed